chore(web): extract app helpers and tidy surfaces
