+++ docs/journal/2026-04-06-agents-lcp-workbench-split.md
@@ -170,6 +170,9 @@ Live verification notes:
 - Follow-up route-shell decomposition extracted the remaining `AgentsRouteShell` prop assembly out of `web/src/app.tsx` into `web/src/components/agents_route_shell_props.ts`, and memoized `AgentsRouteShell`/`AgentsRouteShellView` so unrelated root rerenders stop rebuilding the shell prop bags by default.
 - Follow-up modal-shell cleanup extracted the remaining create-agent / node-section / permission-modal prop assembly into `web/src/components/agents_route_modal_props.ts`, so `web/src/app.tsx` no longer builds those large modal prop bags inline during every root render.
 - Follow-up root-view extraction moved the authenticated agents shell and login-form JSX out of `web/src/app.tsx` into `web/src/components/agents_root_page.tsx`, so the root route file now stays focused on route/state orchestration instead of rendering the entire header/auth/modal tree inline.
+- Follow-up route-selection extraction moved the root/admin/team route classification helpers into `web/src/app_route_selection.ts`, added focused route selection coverage in `web/src/app_route_selection.test.ts`, and converted the bottom of `web/src/app.tsx` onto a single `resolveAppRouteKind(...)` switch so route composition no longer duplicates pathname/auth branching inline.
+- Follow-up helper extraction moved the remaining agents/workbench pure helpers into `web/src/app_agents_helpers.ts` and re-exported them from `web/src/app.tsx`, reducing the root route file from `3300` lines to `3063` while keeping `app.route_auth` / `app.permission_scope` compatibility imports unchanged.
+- The latest local route-selection pass stayed green on `app_route_selection`, `app.route_auth`, `app.permission_scope`, lint, build, and `make build-web`; the live MCP after-check could not be completed because both `/` and `/teams/...` were returning `502 Bad Gateway` during verification, so this pass relies on local evidence rather than a fresh deployed browser regression.
 
 ## Follow-up
